Fixing misc bugs

+++ src/DF2/DF_util.h
@@ -2,7 +2,7 @@
 #include "DF_def.h"
 
 template<size_t N>
-DF<N> DF_index(const DF<N>& in, const std::vector<int>& idxs)
+DF<N> DF_index(const DF<N>& in, const std::vector<size_t>& idxs)
 {
     DF<N> out = Eigen::ArrayXXd::Zero(idxs.size(), N);
     for (size_t i = 0; i < idxs.size(); i++)
